requirements.txt
django>=5.0
django-tenants>=3.6.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
djoser>=2.2.2
django-cors-headers>=4.3.1
google-generativeai>=0.3.0
openai>=1.3.0
anthropic>=0.18.0
requests>=2.31.0
gunicorn>=21.2.0
python-magic>=0.4.27
Pillow>=10.0.0
django-humanize>=0.0.1
